make rm on script itself only if script actually remains.

+++ ipinfo/deb.sh
@@ -9,4 +9,6 @@ rm ipinfo_${VSN}.deb
 echo
 echo 'You can now run `ipinfo`'.
 
-rm $0
+if [ -f "$0" ]; then
+    rm $0
+fi
