+++ ipinfo/macos.sh
@@ -11,4 +11,6 @@ mv ipinfo_${VSN}_${PLAT} /usr/local/bin/ipinfo
 echo
 echo 'You can now run `ipinfo`'.
 
-rm $0
+if [ -f "$0" ]; then
+    rm $0
+fi
